Revert "Merge pull request #2214 from conda/kalefranz/conda-common"
This reverts commit 0629df9e50fee2e4b2dc45a85ab6999603bd854f, reversing
changes made to 8817ff777735aed666ae01f4680c8df05db5d085.
