+++ conda/packup.py
@@ -3,20 +3,20 @@
 #     new code.
 from __future__ import print_function, division, absolute_import
 
-import hashlib
-import json
 import os
 import re
-import shutil
 import sys
+import json
+import shutil
+import hashlib
 import tarfile
 import tempfile
 from os.path import basename, isfile, islink, join
 
 import conda.config as config
 import conda.install as install
-from conda.common.compat import PY3
 from conda.misc import untracked
+from conda.compat import PY3
 
 
 def get_installed_version(prefix, name):
