+++ setup.py
@@ -63,7 +63,6 @@ setup(
     packages=[
         'conda',
         'conda.cli',
-        'conda.common',
         'conda.progressbar'
     ],
     cmdclass={
